plugin/forward: correctly set TLS config per proxy
The TLS config was not correctly propegated to the healthcheck ping,
meaning it would default to UDP on port 853 (e.g.). Fix this and add
some extra testing. Also make the config apply to a single proxy and
not all proxies (some of which may not use TLS)

With this Corefile:

~~~
. {
    log
    errors
    forward . tls://9.9.9.9 8.8.8.8 {
        tls_servername dns.quad9.net
    }
}
~~~

I see this debug scrolling by:

~~~
PRO udp 8.8.8.8:53
[::1]:45932 - [14/Apr/2018:14:37:03 +0100] 25934 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 53.034632ms
PRO tcp-tls 9.9.9.9:853
[::1]:50867 - [14/Apr/2018:14:37:04 +0100] 28900 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 154.203886ms
PRO tcp-tls 9.9.9.9:853
[::1]:49818 - [14/Apr/2018:14:37:06 +0100] 21654 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 14.453526ms
PRO tcp-tls 9.9.9.9:853
[::1]:41386 - [14/Apr/2018:14:37:07 +0100] 36473 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 14.824108ms
PRO tcp-tls 9.9.9.9:853
[::1]:54474 - [14/Apr/2018:14:37:08 +0100] 24185 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 15.387741ms
PRO udp 8.8.8.8:53
[::1]:32794 - [14/Apr/2018:14:37:08 +0100] 13500 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 13.88044ms
~~~
I.e. the right protocol for the right upstream.

With `force_tcp` added to the mix:

~~~
RO tcp 8.8.8.8:53
[::1]:51204 - [14/Apr/2018:14:37:41 +0100] 31503 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 51.63207ms
PRO tcp-tls 9.9.9.9:853
[::1]:50995 - [14/Apr/2018:14:37:42 +0100] 61789 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 158.071374ms
PRO tcp-tls 9.9.9.9:853
[::1]:42286 - [14/Apr/2018:14:37:43 +0100] 7200 "MX IN miek.nl. udp 37 false 4096" NOERROR qr,rd,ra,ad 170 15.982663ms
~~~

TCP instead of UDP for 8.8.8.8, but still TLS for 9.9.9.9

Fixes #1672